+++ src/app.module.ts
@@ -2,6 +2,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UsersModule } from './users/users.module';
 import { AuthModule } from './auth/auth.module';
+import { TransactionsModule } from './transactions/transactions.module';
 
 @Module({
   imports: [
@@ -16,7 +17,8 @@ import { AuthModule } from './auth/auth.module';
       synchronize: true, // TODO solo en dev
     }),
     UsersModule,
-    AuthModule
+    AuthModule,
+    TransactionsModule
   ],
 })
 export class AppModule {}